Move API types into separate files
